# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Vaccine:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      interval:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sort_order:
        type: integer
      species:
        $ref: '#/definitions/model.VaccineSpecies'
      updated_at:
        type: string
    type: object
  model.VaccineSpecies:
    enum:
    - dog
    - cat
    - both
    type: string
    x-enum-varnames:
    - VaccineSpeciesDog
    - VaccineSpeciesCat
    - VaccineSpeciesBoth
...
host: localhost:8080
info:
  contact: {}
  description: 動物病院 電子カルテシステム API（45テーブル・専用マスタテーブル版）
  title: Animal Ekarte API
  version: "2.0"
paths:
...
  /masters/vaccines:
    get:
      description: クリニックに登録されているワクチンの一覧を返す。speciesクエリで対象種絞り込み可能。
      parameters:
      - description: '対象種でフィルタリング（例: dog, cat, both）'
        in: query
        name: species
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Vaccine'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ワクチン一覧取得
      tags:
      - VaccineMasters
    post:
      consumes:
      - application/json
      description: 新規ワクチンを登録する。
      parameters:
      - description: ワクチン情報
        in: body
        name: vaccine
        required: true
        schema:
          $ref: '#/definitions/model.Vaccine'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Vaccine'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ワクチン作成
      tags:
      - VaccineMasters
  /masters/vaccines/{id}:
    delete:
      description: 指定IDのワクチンを削除する。
      parameters:
      - description: ワクチンID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ワクチン削除
      tags:
      - VaccineMasters
    put:
      consumes:
      - application/json
      description: 指定IDのワクチン情報を更新する。
      parameters:
      - description: ワクチンID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新するワクチン情報
        in: body
        name: vaccine
        required: true
        schema:
          $ref: '#/definitions/model.Vaccine'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Vaccine'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ワクチン更新
      tags:
      - VaccineMasters
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"